use killpg to kill workers
`killpg` will kill not just the worker but also any subprocess under it. Currently a check that starts another process (eg. `exasol()`) which runs longer then interval will timeout, which will cause worker monitor to kill worker process but not the hcild. The child will be orphaned and remain in the system indefinitely.