fix: too large captures for opsgenie
we limit the amount of captures being sent to opsgenie to half of
the max payload size - the rest is small and should not exceed 1-2kB
in normal usage. The captures are visible in the alert detail
page also. The include_captures=True is a convenience feature. It is
better to drop the captures than not to send out the page